requirements.txt
# PneumoScan — Python Dependencies
# Compatible with Python 3.13 and TensorFlow 2.20.0
#
# IMPORTANT: numpy 1.x has no pre-built cp313 wheels and requires GCC ≥ 8.4
# to compile from source. Use numpy 2.x which ships ready-to-install cp313
# wheels. TensorFlow 2.16+ supports numpy 2.x.
#
# Keras compatibility: TF 2.16+ ships Keras 3 by default. tf_keras re-exports
# the legacy Keras 2 API that tf-keras-vis requires.

tensorflow==2.20.0
tf_keras==2.20.0

numpy==2.2.4
pandas==2.2.3
scikit-learn==1.6.1
matplotlib==3.10.1
seaborn==0.13.2
opencv-python==4.11.0.86
Pillow==11.2.1
flask==3.1.0
flask-cors==5.0.1
tf-keras-vis==0.8.7
